+++ api/types/restaurant/index.ts
@@ -1,17 +1,1 @@
-interface Restaurant {
-  id: string;
-  name: string;
-  url: string;
-  image_url: string;
-  distance: number;
-  price: string;
-  display_address: string[];
-  rating: number;
-  location: { display_address: string[] };
-  categories: { title: string }[];
-  attributes?: {
-    menu_url?: string;
-  };
-}
-
-export type { Restaurant };
+export type { Restaurant } from '@lunch-roulette/types';
