+++ package.json
@@ -22,6 +22,7 @@
   "author": "Chris Samuel",
   "license": "MIT",
   "dependencies": {
+    "@lunch-roulette/types": "workspace:*",
     "@slack/types": "2.20.1",
     "@slack/web-api": "7.15.0",
     "@vercel/node": "5.6.18",
