+++ packages/types/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@lunch-roulette/types",
+  "version": "1.0.0",
+  "private": true,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts"
+}
