+++ pnpm-workspace.yaml
@@ -1,0 +1,6 @@
+packages:
+  - 'packages/*'
+  - 'web'
+
+allowBuilds:
+  esbuild: true
