+++ vercel.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://openapi.vercel.sh/vercel.json",
-  "buildCommand": "cd web && pnpm install && pnpm build",
+  "buildCommand": "pnpm --filter lunch-roulette-web run build",
   "outputDirectory": "web/dist",
   "functions": {
     "api/**/*": {
