+++ web/package.json
@@ -8,6 +8,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@lunch-roulette/types": "workspace:*",
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-router-dom": "7.6.1"
